Move RearrangeChannels to HAL. Something feels off... Two RearrangeChannels doing nearly the same stuff. Lots of testing is needed.
