+++ modules/core/src/image.cpp
@@ -46,29 +46,6 @@ void Image::Create(const std::vector<int>& dims, DataType elemtype, std::string 
     }
 }
 
-/** @brief Rearrange channels between Images of different DataTypes. */
-template<DataType SDT, DataType DDT>
-struct StructRearrangeImage
-{
-    static void _(const Image& src, Image& dst, const std::vector<int>& bindings)
-    {
-        using dsttype = typename TypeInfo<DDT>::basetype;
-        ConstView<SDT> vsrc(src);
-        View<DDT> vdst(dst);
-        auto id = vdst.Begin();
-
-        for (size_t tmp_pos = 0; tmp_pos < dst.datasize_; tmp_pos += dst.elemsize_, ++id) {
-            int x = static_cast<int>(tmp_pos);
-            int src_pos = 0;
-            for (int i = vsize(dst.dims_) - 1; i >= 0; i--) {
-                src_pos += (x / dst.strides_[i]) * src.strides_[bindings[i]];
-                x %= dst.strides_[i];
-            }
-
-            *id = static_cast<dsttype>(*(vsrc.data_ + src_pos));
-        }
-    }
-};
 
 void RearrangeAndCopy(const Image& src, Image& dst, const std::string& channels, DataType new_type)
 {
@@ -170,8 +147,7 @@ void RearrangeAndCopy(const Image& src, Image& dst, const std::string& channels,
         }
     }
 
-    static constexpr Table2D<StructRearrangeImage> table;
-    table(src.elemtype_, dst.elemtype_)(src, dst, bindings);
+    dst.hal_->RearrangeChannels(src, dst, bindings);
 }
 
 void RearrangeChannels(const Image& src, Image& dst, const std::string& channels)
@@ -216,7 +192,7 @@ void RearrangeChannels(const Image& src, Image& dst, const std::string& channels
             x %= tmp.strides_[i];
         }
 
-        memcpy(tmp.data_ + tmp_pos, src.data_ + src_pos, tmp.elemsize_);
+        tmp.hal_->MemCopy(tmp.data_ + tmp_pos, src.data_ + src_pos, tmp.elemsize_);
     }
 
     // TODO consider spacings
